Set random cuda device only if cuda is requested.
We set the current cuda device to a random one at the start to
distribute between students.  However, not all practicals need a CPU
and some are optional.  This prevented the practical from running
cleanly on other places that do not have a GPU.

This commit moves the selection of cuda device only where use_gpu is
set.