category-recognition-cnn-pytorch: replace requests with urllib.
requests is a 3rd party package while urllib is a part of the Python
standard library.  While requests is very nice, we are not doing
anything nice and it's still a one liner in urllib.  Also, urllib
raises exception if the error code is not success (it was previously
failing silently leading to later issues).  Finally, urllib returns a
stream which PIL handles automatically so no need for io.BytesIO.